Implement Set method for inner node
