+++ pkg/config/nodetreemodel/config.go
@@ -47,8 +47,14 @@ var sources = []model.Source{
 // - contains metadata about known keys, env var support
 type ntmConfig struct {
 	sync.RWMutex
-	root   Node
 	noimpl notImplementedMethods
+
+	// defaults contains the settings with a default value
+	defaults Node
+	// file contains the settings pulled from YAML files
+	file Node
+	// root contains the final configuration, it's the result of merging all other tree by ordre of priority
+	root Node
 
 	envPrefix      string
 	envKeyReplacer *strings.Replacer
@@ -91,29 +97,42 @@ func (c *ntmConfig) getValue(key string) (interface{}, error) {
 	return c.leafAtPath(key).GetAny()
 }
 
-func (c *ntmConfig) setValueSource(key string, newValue interface{}, source model.Source) {
+func (c *ntmConfig) setValueSource(key string, newValue interface{}, source model.Source) { // nolint: unused // TODO fix
 	err := c.leafAtPath(key).SetWithSource(newValue, source)
 	if err != nil {
 		log.Errorf("%s", err)
 	}
 }
 
+func (c *ntmConfig) set(key string, value interface{}, tree Node, source model.Source) (bool, error) {
+	parts := strings.Split(strings.ToLower(key), ",")
+	return tree.Set(parts, value, source)
+}
+
 // Set assigns the newValue to the given key and marks it as originating from the given source
 func (c *ntmConfig) Set(key string, newValue interface{}, source model.Source) {
-	if source == model.SourceDefault {
-		c.SetDefault(key, newValue)
+	var tree Node
+
+	// TODO: have a dedicated mapping in ntmConfig instead of a switch case
+	switch source {
+	case model.SourceDefault:
+		tree = c.defaults
+	case model.SourceFile:
+		tree = c.file
+	}
+
+	c.Lock()
+
+	previousValue, _ := c.getValue(key)
+	_, _ = c.set(key, newValue, tree, source)
+	updated, _ := c.set(key, newValue, c.root, source)
+
+	// if no value has changed we don't notify
+	if !updated || reflect.DeepEqual(previousValue, newValue) {
 		return
 	}
 
-	// modify the config then release the lock to avoid deadlocks while notifying
-	var receivers []model.NotificationReceiver
-	c.Lock()
-	previousValue, _ := c.getValue(key)
-	c.setValueSource(key, newValue, source)
-	if !reflect.DeepEqual(previousValue, newValue) {
-		// if the value has not changed, do not duplicate the slice so that no callback is called
-		receivers = slices.Clone(c.notificationReceivers)
-	}
+	receivers := slices.Clone(c.notificationReceivers)
 	c.Unlock()
 
 	// notifying all receiver about the updated setting
@@ -129,7 +148,10 @@ func (c *ntmConfig) SetWithoutSource(key string, value interface{}) {
 
 // SetDefault assigns the value to the given key using source Default
 func (c *ntmConfig) SetDefault(key string, value interface{}) {
-	c.Set(key, value, model.SourceDefault)
+	c.Lock()
+	defer c.Unlock()
+
+	_, _ = c.set(key, value, c.defaults, model.SourceDefault)
 }
 
 // UnsetForSource unsets a config entry for a given source
